app/requirements.txt
pip
certifi
chardet
Django==2.2
django-extensions
django-paypal
django-redis
django-redis-cache
django-user-agents
django-oauth-toolkit
django-oauth2-provider
geoip2
certbot
django-map-widgets
jsonfield
xlwt
matplotlib
openpyxl
treepoem
django-webpush
#pygraphviz
fcm-django
tqdm
django-progressive-web-app
django-modeltranslation
certbot-django
django-sslify
certbot-dns-ovh
drf-yasg==1.15.0
drfdocs
django-cors-headers==2.5.2
django-rest-swagger==2.2.0
djangorestframework==3.9.2
stripe
PyPDF2
idna
ipaddress
django-letsencrypt
django-sslserver
Markdown
maxminddb
numpy
pandas
oauthlib
Pillow
psycopg2-binary
PyJWT
pygments
markdown
pyyaml
python-dateutil
python-openid
python-social-auth
djangorestframework-gis
pytz
coreapi
PyYAML
redis
requests
requests-oauthlib
#six
social-auth-app-django
social-auth-core
typing
ua-parser
urllib3
user-agents
virtualenv
exifread
django-decorator-include
xlrd
lxml
wget
#torchvision
#==0.4.0
#torch
#==1.2.0
scipy
cupy-cuda90
librosa
#==0.7.1
unidecode
inflect
Cython
#pysndfilesndfile
sndfile
#pysndfile
playsound
pydub
ffmpeg
ffmpy
numba==0.48
#keras
psutil
SpeechRecognition
multiprocess
webrtcvad
kaldi-io
marshmallow
num2words
ruamel.yaml
soundfile
sox
torch
#==1.5.0
torchvision
#==0.6.0
torch-stft
unidecode
nemo-toolkit==0.9.0
nemo-asr==0.9.0
cython
setuptools
llvmlite==0.31.0
IPython
torchaudio